formalized some type conversion to stop error log

+++ OpenGLExperiments/OpenGLExperiments/src/Classes/Manager.cpp
@@ -54,7 +54,7 @@ void Manager::spawnObject(glm::vec3 new_position, glm::vec3 new_velocity, float 
 	position.push_back(new_position);
 	velocity.push_back(new_velocity);
 	halfWidth.push_back(new_radius);
-	mass.push_back(pow(new_radius, 3));
+	mass.push_back(pow(new_radius, 3.0f));
 	restitution.push_back(min_restitution + randomVal(max_restitution - min_restitution, true));
 	count++;
 }
@@ -66,7 +66,7 @@ void Manager::spawnObjects(int n)
 		position.push_back(randomVec3(boundarySize));
 		velocity.push_back(randomVec3(maxSpeed));
 		halfWidth.push_back(randomVal(maxWidth / 2, true));
-		mass.push_back(pow(halfWidth[i],3));
+		mass.push_back(pow(halfWidth[i],3.0f));
 		restitution.push_back(min_restitution + randomVal(max_restitution - min_restitution, true));
 		count++;
 	}
@@ -79,7 +79,7 @@ void Manager::spawnCoplanarObjects(int n)
 		position.push_back(glm::vec3(randomVal(boundarySize), 0, randomVal(boundarySize)));
 		velocity.push_back(glm::vec3(randomVal(maxSpeed), 0, randomVal(maxSpeed)));
 		halfWidth.push_back(randomVal(maxWidth / 2, true));
-		mass.push_back(pow(halfWidth[i], 3));
+		mass.push_back(pow(halfWidth[i], 3.0f));
 		restitution.push_back(min_restitution + randomVal(max_restitution - min_restitution, true));
 		count++;
 	}
@@ -88,11 +88,11 @@ void Manager::spawnCoplanarObjects(int n)
 void Manager::generateBuffers()
 {
 	glGenBuffers(1, &VBO_pos);
-	object->setVertices(&position[0], position.size() * 3, object->_VAO, VBO_pos);
+	object->setVertices(&position[0], (int)position.size() * 3, object->_VAO, VBO_pos);
 	object->setAttributes(1, 3, GL_FLOAT, GL_FALSE, 3 * sizeof(float), (void*)(0));
 	glVertexAttribDivisor(1, 1);
 	glGenBuffers(1, &VBO_radius);
-	object->setVertices(&halfWidth[0], halfWidth.size(), object->_VAO, VBO_radius);
+	object->setVertices(&halfWidth[0], (int)halfWidth.size(), object->_VAO, VBO_radius);
 	object->setAttributes(2, 1, GL_FLOAT, GL_FALSE,sizeof(float), (void*)(0));
 	glVertexAttribDivisor(2, 1);
 }
@@ -104,7 +104,7 @@ void Manager::drawObjects(Camera camera)
 	object->shader->setMat4("VP", VP);
 	object->shader->setVec3("viewPos", camera._position);
 	glBindVertexArray(object->_VAO);
-	object->setVertices(&position[0], position.size() * 3, object->_VAO, VBO_pos);
-	object->setVertices(&halfWidth[0], halfWidth.size(), object->_VAO, VBO_radius);
+	object->setVertices(&position[0], (int)position.size() * 3, object->_VAO, VBO_pos);
+	object->setVertices(&halfWidth[0], (int)halfWidth.size(), object->_VAO, VBO_radius);
 	glDrawElementsInstanced(GL_TRIANGLES, object->indexCount,  GL_UNSIGNED_INT,0, count);
 }
